[FLINK-31114][runtime] Refactor the package path of ForwardGroup
